# gamification-impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamification API
  description: description
security:
  - apiKeyAuth: []
paths:
  /applications:
    post:
      description: create a new application
      operationId: createApplication
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/NewApplication'
      responses:
        '201':
          description: new application created
          content:
            'application/json':
              schema:
                type: string
  /applications/{name}:
    get:
      description: get one application, by name
      operationId: getApplication
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Application'
        '404':
          description: if application name is unknown
  /badges:
    post:
      description: create badge for the website
      operationId: createBadge
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                type: string
        '400':
          description: parameters are empty or null
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '409':
          description: badge already exists
    get:
      description: get the list of all badges
      operationId: getBadges
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Badge'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
  /badges/{name}:
    get:
      description: get one badge, by name
      operationId: getBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: badge not found
    put:
      description: update a badge, by name
      operationId: putBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
        - in: query
          name: usable
          schema:
            type: boolean
            default: true
          description: if the badge has to be desactivated pass this value with false
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Badge'
        description: pass an empty body if you would like to only modify the usable value
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Badge'
        '400':
          description: request body was incorrect
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '409':
          description: badge's name conflict

    delete:
      description: delete the specified badge
      operationId: removeBadge
      parameters:
        - in: path
          name: name
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Badge Removed
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Badge not found
  /events:
    post:
      description: create a user event
      operationId: createEvent
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: created
          content:
            'application/json':
              schema:
                type: string
        '401':
          $ref: "#/components/responses/UnauthorizedError"
    get:
      description: get the list of all events
      operationId: getEvents
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventInfo'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
  /events/{id}:
    get:
      description: get one event, by id
      operationId: getEvent
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Event'
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '404':
          description: Event not found
...
components:
  securitySchemes:
    #    bearerAuth:            # arbitrary name for the security scheme
    #      type: http
    #      scheme: bearer       # warning: it HAS TO be a lowercase b for the swagger ui version used in springfox
    #      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
...
  responses:
    UnauthorizedError:
      description: API key is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    Event:
      type: object
      properties:
        userAppId:
          type: string
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
        eventProperties:
          type: object
    EventInfo:
      type: object
      properties:
        id:
          type: integer
        userAppId:
          type: string
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
    Badge:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    NewApplication:
      type: object
      properties:
        name:
          type: string
    Application:
      type: object
      properties:
        apiKey:
          type: string
        name:
          type: string
...